requirements.txt

google-generativeai
groq
edge-tts
moviepy
sqlalchemy
google-auth-oauthlib
google-api-python-client
requests
wikipedia
pillow
numpy
python-dotenv
